backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm_evaluation_framework"
version = "0.1.0"
description = "A framework for evaluating LLM-based micro-agents"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.11.0",
    "asyncpg>=0.27.0",
    "httpx>=0.24.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.6",
    "celery>=5.3.1",
    "redis>=4.6.0",
    "mlflow>=2.7.0",
    "wandb>=0.15.5",
    "authlib>=1.2.0",
    "itsdangerous>=2.1.2",
    "ragas>=0.0.18",
    "deepeval>=0.20.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pytest>=7.3.1",
    "pytest-asyncio>=0.21.0",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "jwt (>=1.3.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "ruff>=0.0.270",
    "pre-commit>=3.3.3",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
